add new room name generator

+++ headers/room.h
@@ -9,19 +9,19 @@ class Room
     private:
         uint16_t roomNumber;
         std::string roomName;
-        uint32_t roomDescLength;
+        uint16_t roomDescLength;
         std::string roomDesc;
-        std::vector<uint8_t> connectedRoomNums;
+        std::vector<uint16_t> connectedRoomNums;
         std::vector<Baddie*> baddieList;
         std::vector<Player*> playerList;
         std::mutex innerlock;
 
     public:
-        Room(uint16_t,std::string,uint32_t,std::string);
+        Room(uint16_t,std::string,uint16_t,std::string);
         ~Room();
         void lock();
         void unlock();
-        void setConnectedRooms(uint8_t);
+        void setConnectedRooms(uint16_t);
         void addBaddie(Baddie);
         void addPlayer(Player*);
         int searchPlayer(std::string const&);
